+++ aisiki/services/__init__.py
@@ -1,4 +1,4 @@
+from . import aisiki_base_rest
 from . import orderingapp
-from . import authenticate
 from . import saleforce
-from . import product
+# from . import product
